Added support for building 3DS/CIA files
There is some pretty invasive changes. The major one is the removal of
the old Makefiles, substitute by Steveice10/buildtools. This made it
much easier to add support for 3DS/CIA files. Had to rename Makefiles
because buildtools thought that Makefile.3ds was a target build too.

3DS/CIA files need a banner.png image and audio.wav, together with an
icon.png file. So I created a banner.png image based on this one:
https://thenounproject.com/term/folder/1249/. This file is in Creative
Commons 3.0 Attribution license, so I added proper credits in README.md
to be compatible with the license. The audio.wav image is ripped of from
Steveice10/FBI, however it can be changed easily if this is needed. I
updated the icon.png file too, to better match the banner.png image.